+++ all/native/rastertiles/TerrariumElevationDataDecoder.cpp
@@ -12,7 +12,7 @@ namespace carto
     {
     }
 
-    std::array<float, 4> TerrariumElevationDataDecoder::getColorComponentCoefficients() const
+    std::array<double, 4> TerrariumElevationDataDecoder::getColorComponentCoefficients() const
     {
         return COMPONENTS;
     }
@@ -22,6 +22,6 @@ namespace carto
         return SCALES;
     }
 
-    const std::array<float, 4> TerrariumElevationDataDecoder::COMPONENTS = std::array<float, 4>{256.0f, 1.0f, 1.0f / 256, -32768.0f};
+    const std::array<double, 4> TerrariumElevationDataDecoder::COMPONENTS = std::array<double, 4>{256.0f, 1.0f, 1.0f / 256, -32768.0f};
     const std::array<float, 4> TerrariumElevationDataDecoder::SCALES = std::array<float, 4>{256.0f, 1.0f, 1.0f / 256, -32768.0f};
 } // namespace carto
